global change event structure
