+++ public/events/entity/import.js
@@ -1,0 +1,13 @@
+import {nameFieldEvents} from './name/name-events.js';
+import {getPhoneFieldEvents} from './phone/phone-events.js';
+import {
+    singlePasswordFieldEvents,
+    doublePasswordFieldEvents
+} from './password/password-events.js';
+
+export const Event = {
+    nameFieldEvents,
+    getPhoneFieldEvents,
+    singlePasswordFieldEvents,
+    doublePasswordFieldEvents,
+};
