+++ public/events/entity/password/password-events.js
@@ -1,0 +1,40 @@
+import * as VALIDATION from "./password-src.js";
+import * as FORM from '../../common/status-form.js';
+import * as CONFIG from '../../common/config.js';
+
+export const singlePasswordFieldEvents = [
+    {
+        type: 'change',
+        /**
+         * @function Осуществляет проверку пароля на валидность в форме авторизации.
+         *      А также производит запись статуса в statusForm данной страницы.
+         *      statusForm редназначен для быстрой проверки формы на валидность перед отправкой.
+         * @param {Object} app - Объект приложения.
+         * @param {Event} e - Событие.
+         */
+        listener(app, e) {
+            FORM.inputDataManager(
+                e, 'loginPassword', FORM.statusLoginForm,
+                CONFIG.Regex.password, CONFIG.ErrorMsg.errorPassword
+            );
+        }
+    }
+];
+
+export const doublePasswordFieldEvents = [
+    {
+        type: 'change',
+        /**
+         * @function Осуществляет проверку паролей на валидность и идентичность в форме авторизации.
+         *      А также производит запись статуса в statusForm данной страницы.
+         *      statusForm редназначен для быстрой проверки формы на валидность перед отправкой.
+         * @param {Object} app - Объект приложения.
+         * @param {Event} e - Событие.
+         */
+        listener(app, e) {
+            VALIDATION.passwordController(
+                e, FORM.statusRegisterForm
+            );
+        }
+    }
+];
